refractors: unify path normalization util
