+++ core/module_specifier.rs
@@ -1,8 +1,7 @@
+use crate::utils::normalize_path;
 use std::env::current_dir;
 use std::error::Error;
 use std::fmt;
-use std::path::Component;
-use std::path::Path;
 use std::path::PathBuf;
 use url::ParseError;
 use url::Url;
@@ -206,42 +205,10 @@ impl PartialEq<String> for ModuleSpecifier {
   }
 }
 
-/// Normalize all itermediate components of the path (ie. remove "./" and "../" components).
-/// Similar to `fs::canonicalize()` but doesn't resolve symlinks.
-///
-/// Taken from Cargo
-/// https://github.com/rust-lang/cargo/blob/af307a38c20a753ec60f0ad18be5abed3db3c9ac/src/cargo/util/paths.rs#L60-L85
-pub fn normalize_path(path: &Path) -> PathBuf {
-  let mut components = path.components().peekable();
-  let mut ret =
-    if let Some(c @ Component::Prefix(..)) = components.peek().cloned() {
-      components.next();
-      PathBuf::from(c.as_os_str())
-    } else {
-      PathBuf::new()
-    };
-
-  for component in components {
-    match component {
-      Component::Prefix(..) => unreachable!(),
-      Component::RootDir => {
-        ret.push(component.as_os_str());
-      }
-      Component::CurDir => {}
-      Component::ParentDir => {
-        ret.pop();
-      }
-      Component::Normal(c) => {
-        ret.push(c);
-      }
-    }
-  }
-  ret
-}
-
 #[cfg(test)]
 mod tests {
   use super::*;
+  use std::path::Path;
 
   #[test]
   fn test_resolve_import() {
